backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
camelot-py
python-multipart
